Add light regulator service

+++ include/domain/Regulator.h
@@ -2,10 +2,11 @@
     #define REGULATOR_H
 
     #include "domain/Device.h"
+    #include <Arduino.h>
 
     class Regulator : public Device {
         public:
-            virtual void setLevel(uint8_t level) = 0;
+            virtual void setLevel(uint16_t level) = 0;
     };
 
 #endif // REGULATOR_H
