+++ src/infrastructure/regulators/MiddleRegulator.cpp
@@ -4,7 +4,7 @@ MiddleRegulator::MiddleRegulator(int pin) : pin(pin) {
     pinMode(pin, OUTPUT);
 }
 
-void MiddleRegulator::setLevel(uint8_t level) {
+void MiddleRegulator::setLevel(uint16_t level) {
     if (level > 1023) {
         level = 1023;
     }
